@@ -1,16 +1,52 @@
 module github.com/go-pkgz/auth/_example
 
-go 1.15
+go 1.21
 
 replace github.com/go-pkgz/auth => ../
 
 require (
-	github.com/go-chi/chi/v5 v5.0.7
-	github.com/go-oauth2/oauth2/v4 v4.5.1
-	github.com/go-pkgz/auth v1.18.0
-	github.com/go-pkgz/lgr v0.10.4
-	github.com/go-pkgz/rest v1.16.0
+	github.com/go-chi/chi/v5 v5.0.10
+	github.com/go-oauth2/oauth2/v4 v4.5.2
+	github.com/go-pkgz/auth v1.22.0
+	github.com/go-pkgz/lgr v0.11.1
+	github.com/go-pkgz/rest v1.17.0
 	github.com/golang-jwt/jwt v3.2.2+incompatible
-	github.com/tidwall/buntdb v1.2.8 // indirect
-	golang.org/x/oauth2 v0.0.0-20220411215720-9780585627b5
+	golang.org/x/oauth2 v0.13.0
+)
+
+require (
+	cloud.google.com/go/compute v1.23.1 // indirect
+	cloud.google.com/go/compute/metadata v0.2.3 // indirect
+	github.com/MicahParks/keyfunc v1.1.0 // indirect
+	github.com/dghubble/oauth1 v0.7.2 // indirect
+	github.com/go-pkgz/repeater v1.1.3 // indirect
+	github.com/golang-jwt/jwt/v4 v4.4.1 // indirect
+	github.com/golang/protobuf v1.5.3 // indirect
+	github.com/golang/snappy v0.0.4 // indirect
+	github.com/google/uuid v1.3.1 // indirect
+	github.com/klauspost/compress v1.17.1 // indirect
+	github.com/montanaflynn/stats v0.7.1 // indirect
+	github.com/nullrocks/identicon v0.0.0-20180626043057-7875f45b0022 // indirect
+	github.com/pkg/errors v0.9.1 // indirect
+	github.com/tidwall/btree v1.7.0 // indirect
+	github.com/tidwall/buntdb v1.3.0 // indirect
+	github.com/tidwall/gjson v1.17.0 // indirect
+	github.com/tidwall/grect v0.1.4 // indirect
+	github.com/tidwall/match v1.1.1 // indirect
+	github.com/tidwall/pretty v1.2.1 // indirect
+	github.com/tidwall/rtred v0.1.2 // indirect
+	github.com/tidwall/tinyqueue v0.1.1 // indirect
+	github.com/xdg-go/pbkdf2 v1.0.0 // indirect
+	github.com/xdg-go/scram v1.1.2 // indirect
+	github.com/xdg-go/stringprep v1.0.4 // indirect
+	github.com/youmark/pkcs8 v0.0.0-20201027041543-1326539a0a0a // indirect
+	go.etcd.io/bbolt v1.3.7 // indirect
+	go.mongodb.org/mongo-driver v1.12.1 // indirect
+	golang.org/x/crypto v0.14.0 // indirect
+	golang.org/x/image v0.13.0 // indirect
+	golang.org/x/sync v0.4.0 // indirect
+	golang.org/x/sys v0.13.0 // indirect
+	golang.org/x/text v0.13.0 // indirect
+	google.golang.org/appengine v1.6.8 // indirect
+	google.golang.org/protobuf v1.31.0 // indirect
 )
