@@ -15,18 +15,18 @@ func NewNoOp() *NoOp { return &NoOp{} }
 func (s *NoOp) String() string { return "" }
 
 // Put is a NoOp implementation
-func (s *NoOp) Put(userID string, reader io.Reader) (avatarID string, err error) { return "", nil }
+func (s *NoOp) Put(string, io.Reader) (avatarID string, err error) { return "", nil }
 
 // Get is a NoOp implementation
-func (s *NoOp) Get(avatarID string) (reader io.ReadCloser, size int, err error) {
+func (s *NoOp) Get(string) (reader io.ReadCloser, size int, err error) {
 	return io.NopCloser(bytes.NewBuffer([]byte(""))), 0, nil
 }
 
 // ID is a NoOp implementation
-func (s *NoOp) ID(avatarID string) (id string) { return "" }
+func (s *NoOp) ID(string) (id string) { return "" }
 
 // Remove is a NoOp implementation
-func (s *NoOp) Remove(avatarID string) error { return nil }
+func (s *NoOp) Remove(string) error { return nil }
 
 // List is a NoOp implementation
 func (s *NoOp) List() (ids []string, err error) { return nil, nil }
