@@ -22,13 +22,14 @@ import (
 
 // CustomHandlerOpt are options to initialize a handler for oauth2 server
 type CustomHandlerOpt struct {
-	Endpoint  oauth2.Endpoint
-	InfoURL   string
-	JwksURL   string
-	LogoutURL string
-	MapUserFn func(UserData, []byte) token.User
-	Scopes    []string
+	Endpoint          oauth2.Endpoint
+	InfoURL           string
+	MapUserFn         func(UserData, []byte) token.User
+	BearerTokenHookFn BearerTokenHook
+	Scopes            []string
 
+	JwksURL           string
+	LogoutURL         string
 	RefreshTokenStore RefreshTokenStore
 }
 
@@ -92,7 +93,8 @@ func (c *CustomServer) Run(ctx context.Context) {
 	}
 
 	c.httpServer = &http.Server{
-		Addr: fmt.Sprintf(":%s", port),
+		Addr:              fmt.Sprintf(":%s", port),
+		ReadHeaderTimeout: 5 * time.Second,
 		Handler: http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 			switch {
 			case strings.HasSuffix(r.URL.Path, "/authorize"):
@@ -138,7 +140,7 @@ func (c *CustomServer) handleAuthorize(w http.ResponseWriter, r *http.Request) {
 				return
 			}
 
-			formData := struct{ Query string }{Query: r.URL.RawQuery}
+			formData := struct{ Query template.URL }{Query: template.URL(r.URL.RawQuery)} //nolint:gosec // query is safe
 
 			if err := userLoginTmpl.Execute(w, formData); err != nil {
 				c.Logf("[WARN] can't write, %s", err)
@@ -211,13 +213,15 @@ func (c *CustomServer) Shutdown() {
 // NewCustom creates a handler for go-oauth2/oauth2 server
 func NewCustom(name string, p Params, copts CustomHandlerOpt) Oauth2Handler {
 	return initOauth2Handler(p, Oauth2Handler{
-		name:              name,
-		endpoint:          copts.Endpoint,
-		scopes:            copts.Scopes,
-		infoURL:           copts.InfoURL,
+		name:            name,
+		endpoint:        copts.Endpoint,
+		scopes:          copts.Scopes,
+		infoURL:         copts.InfoURL,
+		mapUser:         copts.MapUserFn,
+		bearerTokenHook: copts.BearerTokenHookFn,
+
 		jwksURL:           copts.JwksURL,
 		logoutURL:         copts.LogoutURL,
-		mapUser:           copts.MapUserFn,
 		refreshTokenStore: copts.RefreshTokenStore,
 	})
 }
