@@ -23,9 +23,21 @@ var testJwtValid = "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJhdWQiOiJ0ZXN0X3N5cyI
 
 var days31 = time.Hour * 24 * 31
 
+type rememberLastBearerTokenHook struct {
+	LastProviderName string       `json:"lastProviderName"`
+	LastUser         token.User   `json:"lastUser"`
+	LastToken        oauth2.Token `json:"lastToken"`
+}
+
+func (h *rememberLastBearerTokenHook) hook(s string, user token.User, o oauth2.Token) {
+	h.LastProviderName = s
+	h.LastUser = user
+	h.LastToken = o
+}
+
 func TestOauth2Login(t *testing.T) {
 
-	teardown := prepOauth2Test(t, 8981, 8982)
+	teardown := prepOauth2Test(t, 8981, 8982, nil)
 	defer teardown()
 
 	jar, err := cookiejar.New(nil)
@@ -37,7 +49,7 @@ func TestOauth2Login(t *testing.T) {
 	require.Nil(t, err)
 	assert.Equal(t, 200, resp.StatusCode)
 	body, err := io.ReadAll(resp.Body)
-	assert.Nil(t, err)
+	assert.NoError(t, err)
 	t.Logf("resp %s", string(body))
 	t.Logf("headers: %+v", resp.Header)
 
@@ -50,7 +62,7 @@ func TestOauth2Login(t *testing.T) {
 
 	u := token.User{}
 	err = json.Unmarshal(body, &u)
-	assert.Nil(t, err)
+	assert.NoError(t, err)
 	assert.Equal(t, token.User{Name: "blah", ID: "mock_myuser1", Picture: "http://example.com/custom.png", IP: ""}, u)
 
 	tk := resp.Cookies()[0].Value
@@ -65,20 +77,49 @@ func TestOauth2Login(t *testing.T) {
 
 	// check admin user
 	resp, err = client.Get("http://localhost:8981/login?site=remark")
-	assert.Nil(t, err)
+	assert.NoError(t, err)
 	assert.Equal(t, 200, resp.StatusCode)
 	body, err = io.ReadAll(resp.Body)
-	assert.Nil(t, err)
+	assert.NoError(t, err)
 	u = token.User{}
 	err = json.Unmarshal(body, &u)
-	assert.Nil(t, err)
+	assert.NoError(t, err)
 	assert.Equal(t, token.User{Name: "blah", ID: "mock_myuser2", Picture: "http://example.com/ava12345.png",
 		Attributes: map[string]interface{}{"admin": true}}, u)
 }
 
+func TestOauth2LoginBearerTokenHook(t *testing.T) {
+
+	btHook := rememberLastBearerTokenHook{}
+	teardown := prepOauth2Test(t, 8981, 8982, btHook.hook)
+	defer teardown()
+
+	jar, err := cookiejar.New(nil)
+	require.Nil(t, err)
+	client := &http.Client{Jar: jar, Timeout: 5 * time.Second}
+
+	// check non-admin, permanent
+	resp, err := client.Get("http://localhost:8981/login?site=remark")
+	require.Nil(t, err)
+	assert.Equal(t, 200, resp.StatusCode)
+
+	assert.Equal(t, "mock", btHook.LastProviderName)
+	assert.Equal(t, "mock_myuser1", btHook.LastUser.ID)
+	assert.Equal(t, "MTQ0NjJkZmQ5OTM2NDE1ZTZjNGZmZjI3", btHook.LastToken.AccessToken)
+
+	// check admin user
+	resp, err = client.Get("http://localhost:8981/login?site=remark")
+	assert.NoError(t, err)
+	assert.Equal(t, 200, resp.StatusCode)
+
+	assert.Equal(t, "mock", btHook.LastProviderName)
+	assert.Equal(t, "mock_myuser2", btHook.LastUser.ID)
+	assert.Equal(t, "MTQ0NjJkZmQ5OTM2NDE1ZTZjNGZmZjI3", btHook.LastToken.AccessToken)
+}
+
 func TestOauth2LoginSessionOnly(t *testing.T) {
 
-	teardown := prepOauth2Test(t, 8981, 8982)
+	teardown := prepOauth2Test(t, 8981, 8982, nil)
 	defer teardown()
 
 	jar, err := cookiejar.New(nil)
@@ -112,7 +153,7 @@ func TestOauth2LoginSessionOnly(t *testing.T) {
 
 func TestOauth2LoginNoAva(t *testing.T) {
 
-	teardown := prepOauth2Test(t, 8981, 8982)
+	teardown := prepOauth2Test(t, 8981, 8982, nil)
 	defer teardown()
 
 	jar, err := cookiejar.New(nil)
@@ -147,7 +188,7 @@ func TestOauth2LoginNoAva(t *testing.T) {
 
 func TestOauth2Logout(t *testing.T) {
 
-	teardown := prepOauth2Test(t, 8691, 8692)
+	teardown := prepOauth2Test(t, 8691, 8692, nil)
 	defer teardown()
 
 	jar, err := cookiejar.New(nil)
@@ -178,13 +219,13 @@ func TestOauth2Logout(t *testing.T) {
 
 func TestOauth2LogoutWithURI(t *testing.T) {
 	expectedLogoutURL := fmt.Sprintf("http://localhost:%d/logout?client_id=cid&logout_uri=http://localhost:%d/home", 8692, 8691)
-	teardown := prepOauth2TestWithConfig(t, 8691, 8692, func(handler *Oauth2Handler) {
+	teardown := prepOauth2TestWithConfig(t, 8691, 8692, nil, func(handler *Oauth2Handler) {
 		handler.logoutURL = expectedLogoutURL
 	})
 	defer teardown()
 
 	jar, err := cookiejar.New(nil)
-	require.Nil(t, err)
+	require.NoError(t, err)
 	client := &http.Client{Jar: jar, Timeout: 5 * time.Second}
 
 	req, err := http.NewRequest("GET", "http://localhost:8691/logout", http.NoBody)
@@ -193,7 +234,7 @@ func TestOauth2LogoutWithURI(t *testing.T) {
 	req.AddCookie(&http.Cookie{Name: "JWT", Value: testJwtValid, HttpOnly: true, Path: "/", MaxAge: expiration, Secure: false})
 	req.Header.Add("X-XSRF-TOKEN", "random id")
 	resp, err := client.Do(req)
-	require.Nil(t, err)
+	require.NoError(t, err)
 	require.Equal(t, 200, resp.StatusCode)
 
 	require.Equal(t, "http://localhost:8691/logout", resp.Request.URL.String())
@@ -219,7 +260,7 @@ func TestOauth2InitProvider(t *testing.T) {
 }
 
 func TestOauth2InvalidHandler(t *testing.T) {
-	teardown := prepOauth2Test(t, 8691, 8692)
+	teardown := prepOauth2Test(t, 8691, 8692, nil)
 	defer teardown()
 
 	client := &http.Client{Timeout: 5 * time.Second}
@@ -250,14 +291,15 @@ func TestMakeRedirURL(t *testing.T) {
 }
 
 func TestOauth2Refresh(t *testing.T) {
-	teardown := prepOauth2Test(t, 8981, 8982)
+	teardown := prepOauth2Test(t, 8981, 8982, nil)
 	defer teardown()
 
 	jar, err := cookiejar.New(nil)
-	require.Nil(t, err)
+	require.NoError(t, err)
 	client := &http.Client{Jar: jar, Timeout: 5 * time.Second}
 
 	resp, err := client.Get("http://localhost:8981/login?site=remark&session=1")
+	require.NoError(t, err)
 	claims := token.Claims{
 		User: &token.User{
 			ID: "mock_myuser1",
@@ -280,7 +322,7 @@ func TestOauth2Refresh(t *testing.T) {
 
 type memRefreshTokenStore map[string]string
 
-var mockRefreshToken = "IwOGYzYTlmM2YxOTQ5MGE3YmNmMDFkNTVk"
+var mockRefreshToken = "IwOGYzYTlmM2YxOTQ5MGE3YmNmMDFkNTVk" //nolint gosec // test credentials
 var mockRefreshTokenStore = memRefreshTokenStore{}
 
 func (m memRefreshTokenStore) Save(t string, claims token.Claims) error {
@@ -294,11 +336,11 @@ func (m memRefreshTokenStore) Load(claims token.Claims) (string, error) {
 
 var provider Oauth2Handler
 
-func prepOauth2Test(t *testing.T, loginPort, authPort int) func() {
-	return prepOauth2TestWithConfig(t, loginPort, authPort, nil)
+func prepOauth2Test(t *testing.T, loginPort, authPort int, btHook BearerTokenHook) func() {
+	return prepOauth2TestWithConfig(t, loginPort, authPort, btHook, nil)
 }
 
-func prepOauth2TestWithConfig(t *testing.T, loginPort, authPort int, config func(handler *Oauth2Handler)) func() {
+func prepOauth2TestWithConfig(t *testing.T, loginPort, authPort int, btHook BearerTokenHook, config func(handler *Oauth2Handler)) func() {
 	provider = Oauth2Handler{
 		name: "mock",
 		endpoint: oauth2.Endpoint{
@@ -315,6 +357,8 @@ func prepOauth2TestWithConfig(t *testing.T, loginPort, authPort int, config func
 			}
 			return userInfo
 		},
+		bearerTokenHook: btHook,
+
 		refreshTokenStore: mockRefreshTokenStore,
 	}
 
@@ -345,13 +389,12 @@ func prepOauth2TestWithConfig(t *testing.T, loginPort, authPort int, config func
 	provider = initOauth2Handler(params, provider)
 	svc := Service{Provider: provider}
 
-	ts := &http.Server{Addr: fmt.Sprintf(":%d", loginPort), Handler: http.HandlerFunc(svc.Handler)}
+	ts := &http.Server{Addr: fmt.Sprintf(":%d", loginPort), Handler: http.HandlerFunc(svc.Handler)} //nolint:gosec
 
 	count := 0
 	useIds := []string{"myuser1", "myuser2"} // user for first and second calls
 
-	//nolint dupl
-	oauth := &http.Server{
+	oauth := &http.Server{ //nolint:gosec
 		Addr: fmt.Sprintf(":%d", authPort),
 		Handler: http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 			log.Printf("[MOCK OAUTH] request %s %s %+v", r.Method, r.URL, r.Header)
@@ -406,7 +449,7 @@ func mockKeyStore(string) (string, error) { return "12345", nil }
 
 type mockAvatarSaver struct{}
 
-func (m *mockAvatarSaver) Put(u token.User, client *http.Client) (avatarURL string, err error) {
+func (m *mockAvatarSaver) Put(u token.User, _ *http.Client) (avatarURL string, err error) {
 	if u.Picture != "" {
 		return "http://example.com/ava12345.png", nil
 	}
